Rename library to `aiotruenas_client`
To be available in pypi as `aiotruenas-client`.

Addresses #4
