+++ aiotruenas_client/websockets/machine.py
